# InputFile determines huge sections automatically