--- a/src/core/io/src/4C_io_elementreader.cpp
+++ b/src/core/io/src/4C_io_elementreader.cpp
@@ -128,7 +128,7 @@ std::pair<int, std::vector<int>> Core::IO::ElementReader::get_element_size_and_i
   // all reading is done on proc 0
   if (Core::Communication::my_mpi_rank(comm_) == 0)
   {
-    for (const auto& element_line : input_.lines_in_section(sectionname_))
+    for (const auto& element_line : input_.lines_in_section_rank_0_only(sectionname_))
     {
       std::istringstream t{std::string{element_line}};
       int elenumber;
@@ -178,7 +178,7 @@ void Core::IO::ElementReader::get_and_distribute_elements(const int nblock, cons
     int bcount = 0;
     int block = 0;
 
-    for (const auto& element_line : input_.lines_in_section(sectionname_))
+    for (const auto& element_line : input_.lines_in_section_rank_0_only(sectionname_))
     {
       ValueParser parser{element_line, {.user_scope_message = "While reading element line: "}};
       const int elenumber = parser.read<int>() - 1;