This PR fixes a problem where huge yaml files containing a huge discretization cannot be read effectively when using many MPI ranks. Input is now read like this, independent of the underlying format:
- Read the complete input file (+ any includes) on rank 0
- Determine huge sections (more than 10,000 entries)
- Communicate everything except the huge sections to all ranks.
- If a huge section is requested on a rank > 0, communicate it then. This means that `lines_in_section` must be a collective call. To support reading on rank 0 only, we now have `lines_in_section_rank_0_only`. Note that this distinction existed before as well but is now made explicit in the function names.

This implementation is much simpler than the on-the-fly reading we had previously. It seems reasonable that the input file must not be so large, that it cannot even fit in memory at once. Note that all of this is temporary anyway until we resolve #60.

As a small test, I checked the total wall time for input (also include creating the discretization etc.) for an example with 1e6 elements (200 MB file size):

- old implementation, dat: around 60s
- old implementation, yaml: around 100s (not relevant because this would have consumed a large amount of memory)
- new implementation, dat: around 30s
- new implementation, yaml: around 50s

Note that yaml is slower because we currently convert the yaml file into a dat-like representation internally. This will change. The message here is not that yaml is inherently slower. I just wanted to ensure that the new implementation is faster than the old one.